mailbox: add TXDONE_DIRECT
Add TXDONE_DIRECT for mailbox controllers that can very quickly acknowledge
transmission of messages because they are e.g. backed by a hardware FIFO.

The mailbox controller on the Apple M1 SoC has a hardware fifo that is
16 messages deep. In the common case this FIFO will always be empty
when a message is sent. With TXDONE_DIRECT those messages can be
acknowledge immediately and there's no need to enable the interrupt
which would directly fire. In the unlikely case that the FIFO is
currently full the controller can still fall back to TXDONE_BY_IRQ.

Signed-off-by: Sven Peter <sven@svenpeter.dev>