ASoC: tas2764: ensure registers are restored on resume
When these codecs are put into hardware shutdown on suspend,
their register state is lost. This necessitates restoring it
upon resume. This must occur before we try to raise the codec
out of the soft shutdown power state we put it in before suspend,
or the codec will crash.

We currently attempt to raise the power state before
restoring the config registers. This sometimes works when
the PWR_CTRL register is the last to be restored, however
more often than not fails as there is no guaranteed order
in which the registers are restored.

Ensure the config registers are restored correctly by calling
regcache_sync() before attempting to activate the codec.

Suggested-by: Martin Povišer <povik+lin@cutebit.org>
Signed-off-by: James Calligeros <jcalligeros99@gmail.com>