ASoC: tas2770: copy working PM ops from tas2764
The PM routine for the TAS2770 and TAS2764 codecs should be identical,
however the PM ops for the TAS2770 code diverges significantly and is
broken in more ways than TAS2764 was.

The suspend function attempts to pull the codec into hardware shutdown
before putting the power register into the software shutdown state,
which it doesn't even attempt to do if the SDZ_GPIO pin is present.
The same is true on resume, where it does not try to restore the
codec into a known power state.

Copy the known-good PM ops from tas2764.c.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>