drm/asahi: Add AsahiData to the platform device drvdata
The newest DRM rust bindings make the device data unavailable via
drm::device::Device<T> to decouple DRM device creation from the device
data as this is required planned driver like Tyr. On the surface it is
required for asahi as well but it should be possible to avoid this.

Signed-off-by: Janne Grunau <j@jannau.net>